separate list page table into child elements (just separating for now)
